requirements.txt
aerich==0.9.2
aiosqlite==0.22.1
anyio==4.12.1
asyncclick==8.3.0.7
certifi==2026.1.4
dictdiffer==0.9.0
dotenv==0.9.9
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iso8601==2.1.0
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg-pool==3.3.0
pypika-tortoise==0.6.3
python-dotenv==1.2.1
python-telegram-bot==22.6
pytz==2025.2
redis==7.1.0
tortoise-orm==0.25.3
typing_extensions==4.15.0
